DEV: Plotting segmentation maps and reference tractogram. Worked on smoothing.
An arbitrary number of segmentation files can be added to the scene.
A reference tractogram can be specified which will determine the static
camera setup and which will be visualised as transparent streamlines.
The smooth camera updates work well now. Smoothness of camera updates
can be specified for fill up phase of the streamline window as well
as during the rest of the visualisation process, respectively.